Update 59901.tokenlist.json
Add Test eMetis and Test seMetis to the listed tokens on Netswap.io on the Metis Sepolia testnet.

On the ENKI.xyz app hovering over the Metis to eMetis exchange button displays the following message: "It’s not yet possible to exchange Metis into Metis currently. Please go to the exchange to exchange it by yourself."

Clicking on the exchange button brings up a dialog box with the following message: "Redemption Coming Soon!
We're working on it! The redemption feature for SEtoken isn't available just yet. In the meantime, feel free to redeem Mtoken on the secondary market.
Swap eMetis/Metis on Netswap"

Currently neither Test eMetis nor Test seMetis are listed on Netswap.io. This proposes to add them to the Netswap.io listings. Links to the .svg logo files in the nass2010/testnet.enkixyz.com/ repository have been provided for each.